Compile the two assertions clang would not let through
The ui job went red on this and nothing else: a two-enumerator enum in C++
holds one bit, so "type < N_FILTERS" and "type < N_TOOLS" are true by
construction, and clang says so as an error under ENABLE_STRICT. Both sit
inside g_assert(), which is why they had never been compiled -- the ui build
turns assertions on now, being a unit test build.

The cast is what this tree already does with the same comparison:
moousertools.cpp has four g_return_val_if_fail((int) type < N_TOOLS) and two
g_assert(type < N_TOOLS), and only the second kind was invisible to the
compiler.

Three enumerators would have been fine -- the range is the smallest bit-field
that holds every value, so MooActionCheckType with three of them can hold 3
and "type < N_ACTION_CHECKS" is a real check. It is two-value enums only, and
C++ only: in C an enum has the range of its underlying type, which is why the
same shape in moowindow.c says nothing.

+++ src/mooedit/mooeditaction.cpp
@@ -31,6 +31,13 @@ typedef enum {
     FILTER_VISIBLE
 } FilterType;
 
+/*
+ * Two enumerators, so in C++ the type holds one bit and "type < N_FILTERS" is
+ * true by construction -- which clang says out loud, as an error under
+ * ENABLE_STRICT. The cast is what the same assertion elsewhere in the tree
+ * uses (moousertools.cpp), and it keeps the line as the statement of intent it
+ * is. It only ever mattered once assertions were compiled in at all.
+ */
 #define N_FILTERS 2
 
 struct MooEditActionPrivate {
@@ -89,7 +96,7 @@ static const char *
 moo_edit_action_get_filter (MooEditAction *action,
                             FilterType     type)
 {
-    g_assert (type < N_FILTERS);
+    g_assert ((int) type < N_FILTERS);
     return action->priv->filters[type] ?
             g_regex_get_pattern (action->priv->filters[type]) : NULL;
 }
@@ -153,7 +160,7 @@ moo_edit_action_set_filter (MooEditAction *action,
 {
     GRegex *tmp;
 
-    g_assert (type < N_FILTERS);
+    g_assert ((int) type < N_FILTERS);
 
     tmp = action->priv->filters[type];
     action->priv->filters[type] = filter ? get_filter_regex (filter) : NULL;
